docstrings on main + florist

+++ src/application/Decoration.java
@@ -4,7 +4,7 @@ public class Decoration extends Product{
 	
     private MaterialType material;
 
-    public Decoration(String name, float price, int quantity, MaterialType material) {
+    public Decoration(String name, double price, int quantity, MaterialType material) {
         super( name,  price,  quantity, ProductType.DECORATION);
         this.material = material;
