+++ src/application/Flower.java
@@ -3,7 +3,7 @@ package application;
 public class Flower extends Product{
     private String colour;
 
-    public Flower(String name, float price, int quantity,String colour) {
+    public Flower(String name, double price, int quantity,String colour) {
         super( name,  price,  quantity, ProductType.FLOWER);
         this.colour = colour;
